pyproject.toml
[project]
name = "autoresearcher"
version = "0.1.0"
description = "Autonomous ML/DL research agent with metric reasoning and custom model hooks"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pydantic>=2.5",
    "pyyaml>=6.0",
    "pandas>=2.0",
    "numpy>=1.24",
    "scikit-learn>=1.3",
    "anthropic>=0.40",
    "httpx>=0.27",
    "langgraph>=0.2",
    "langchain-core>=0.3",
    "ddgs>=9.0",
]

[project.optional-dependencies]
torch = ["torch>=2.0", "torchvision>=0.15"]
nlp = ["transformers>=4.36", "datasets>=2.16"]
dev = ["pytest>=8.0", "ruff>=0.4"]

[tool.ruff]
line-length = 100
